Simplification of the TreeView API
Dropped TreeViewController and merged
it's logic into TreeViewState to make the
code less 'spaghetti'. Instead of the
controller, use a [GlobalKey<TreeViewState>
to control the TreeView for now on.

Added scrollController optional property to
TreeView new and scrolling functionality.

Renamed ScopedTreeNode to TreeNodeScope and
InheritedTreeView to _TreeViewScope.

Dropped reversedSubtreeGenerator as it was
not being used.